Add readline.h to refactor header dependency
