+++ src/exec/exec.c
@@ -6,11 +6,15 @@
 /*   By: yliu <yliu@student.42.jp>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/06 12:54:34 by yliu              #+#    #+#             */
-/*   Updated: 2024/09/06 14:32:48 by yliu             ###   ########.fr       */
+/*   Updated: 2024/09/06 18:36:52 by yliu             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <exec.h>
+#include "exec.h"
+#include <errno.h>
+#include <stdlib.h>
+#include <string.h>
+#include <unistd.h>
 
 static void	execute_command(char *complete_command, char **envp)
 {
